Return Result from EmailAddress::new()
When instantiating an EmailAddress, using new() is the most natural API; it
also allows the user not to care about which parsing algorithm is used under
the covers, providing less friction for the user. However, due to the use
of unwrap() inside new(), invalid email addresses will panic. Instead,
propagate out the Result type from the parser to new()'s caller.